Started fixing the screenshot function.
- D3D8 is a bitch for this, I got everything as it should be.. But it
keeps throwing a Byref Type Mismatch at me from RECT, which is odd
considering the value IS a RECT, I've tried several ways of declaring it
in and outside the sub but can not get it to work as of yet. Will look
into it again when I have the time.